VM::createPageTable() take boot_pgtable_base as an argument

+++ include/boot.hpp
@@ -27,15 +27,15 @@ constexpr int      PARAGRAPH_SIZE      = 16;
 constexpr int      SECT_SIZE           = 512;
 
 constexpr uint64_t REALMODE_IVT_START  = 0x0000'0000;
-constexpr uint32_t BOOT_PARAMS_ADDR    = 0x0001'0000;
-constexpr uint32_t COMMANDLINE_ADDR    = 0x0002'0000;
-constexpr uint32_t BOOT_PAGETABLE_BASE = 0x0003'0000;
-constexpr uint32_t EBDA_START          = 0x0009'fc00;
-constexpr uint32_t VGARAM_START        = 0x000a'0000;
-constexpr uint32_t MBBIOS_START        = 0x000f'0000;
-constexpr uint32_t HIGHMEM_BASE        = 0x0010'0000;
-constexpr uint32_t INITRAMFS_ADDR      = 0x0f00'0000;
-constexpr uint32_t APIC_BASE           = 0xfee0'0000;
+constexpr uint64_t BOOT_PARAMS_ADDR    = 0x0001'0000;
+constexpr uint64_t COMMANDLINE_ADDR    = 0x0002'0000;
+constexpr uint64_t BOOT_PAGETABLE_BASE = 0x0003'0000;
+constexpr uint64_t EBDA_START          = 0x0009'fc00;
+constexpr uint64_t VGARAM_START        = 0x000a'0000;
+constexpr uint64_t MBBIOS_START        = 0x000f'0000;
+constexpr uint64_t HIGHMEM_BASE        = 0x0010'0000;
+constexpr uint64_t INITRAMFS_ADDR      = 0x0f00'0000;
+constexpr uint64_t APIC_BASE           = 0xfee0'0000;
 constexpr uint32_t MBBIOS_SIZE         = 0x000f'ffff;
 
 constexpr int      EBDA_PADDING_SIZE   = 16*3;
